chore(my-diary): set in order project structure and components
ISSUES CLOSED: #59
