+++ practice/my-diary/src/app/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/app",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
