+++ practice/my-diary/src/components/app-header/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/app-header",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
