--- practice/my-diary/src/components/app/App.tsx
@@ -1,25 +1,0 @@
-import './app.css'
-
-import React from 'react'
-
-import AppHeader from '../app-header'
-import PostAddForm from '../post-add-form'
-import PostList from '../post-list'
-import PostStatusFilter from '../post-status-filter'
-import SearchPanel from '../search-panel'
-
-const App = () => {
-  return (
-    <div className="app">
-      <AppHeader />
-      <div className="search-panel d-flex">
-        <SearchPanel />
-        <PostStatusFilter />
-      </div>
-      <PostList />
-      <PostAddForm />
-    </div>
-  )
-}
-
-export default App
