+++ practice/my-diary/src/components/post-add-button/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/post-add-button",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
