+++ practice/my-diary/src/components/post-add-button/src/PostAddButton.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const PostAddButton = () => {
+  return (
+    <button type="submit" className="btn btn-outline-secondary">
+      Add
+    </button>
+  )
+}
+
+export default PostAddButton
