+++ practice/my-diary/src/components/post-add-button/src/index.tsx
@@ -1,0 +1,3 @@
+import PostAddButton from './PostAddButton'
+
+export default PostAddButton
