+++ practice/my-diary/src/components/post-add-form/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/post-add-form",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
