+++ practice/my-diary/src/components/post-add-input/src/PostAddInput.tsx
@@ -1,8 +1,6 @@
-import './post-add-form.css'
-
 import React from 'react'
 
-const PostAddForm = () => {
+const PostAddInput = () => {
   return (
     <form className="bottom-panel d-flex">
       <input
@@ -10,11 +8,8 @@ const PostAddForm = () => {
         placeholder="what are you thinking today"
         className="form-control new-post-label"
       />
-      <button type="submit" className="btn btn-outline-secondary">
-        Добавить
-      </button>
     </form>
   )
 }
 
-export default PostAddForm
+export default PostAddInput
