+++ practice/my-diary/src/components/post-add-input/src/index.tsx
@@ -1,0 +1,3 @@
+import PostAddInput from './PostAddInput'
+
+export default PostAddInput
