+++ practice/my-diary/src/components/post-list-item/src/index.tsx
