+++ practice/my-diary/src/components/post-list/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/post-list",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
