+++ practice/my-diary/src/components/post-list/src/index.tsx
