+++ practice/my-diary/src/components/post-list/src/post-list.css
@@ -1,5 +1,5 @@
 .app-list {
-    margin-top: 50px;
+    margin: 50px 0 50px 0;
 }
 
 .app-list .list-group-item {
