+++ practice/my-diary/src/components/post-status-filter/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/post-status-filter",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
