+++ practice/my-diary/src/components/search-panel-input/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/search-panel-input",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
