+++ practice/my-diary/src/components/search-panel-input/src/SearchPanelInput.tsx
@@ -1,9 +1,7 @@
-import './search-panel.css'
-
 import React from 'react'
 
-const SearchPanel = () => {
+const SearchPanelInput = () => {
   return <input className="form-control search-input" type="text" placeholder="Search by records" />
 }
 
-export default SearchPanel
+export default SearchPanelInput
