+++ practice/my-diary/src/components/search-panel-input/src/index.tsx
@@ -1,0 +1,3 @@
+import SearchPanelInput from './SearchPanelInput'
+
+export default SearchPanelInput
