+++ practice/my-diary/src/components/search-panel/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/search-panel",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
