+++ practice/my-diary/src/components/star-button/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@components/star-button",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.tsx"
+}
