+++ practice/my-diary/src/components/star-button/src/StarButton.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const StarButton = () => {
+  return (
+    <button className="btn-star btn-sm" type="button">
+      <i className="fa fa-star"></i>
+    </button>
+  )
+}
+
+export default StarButton
