+++ practice/my-diary/src/components/star-button/src/index.tsx
@@ -1,0 +1,3 @@
+import StarButton from './StarButton'
+
+export default StarButton
