+++ practice/my-diary/src/components/trash-button/.eslintrc
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "import/no-extraneous-dependencies": 0
+  }
+}
