+++ practice/my-diary/src/components/trash-button/src/TrashButton.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const TrashButton = () => {
+  return (
+    <button className="btn-trash btn-sm" type="button">
+      <i className="fa fa-trash-o"></i>
+    </button>
+  )
+}
+
+export default TrashButton
