+++ practice/my-diary/src/components/trash-button/src/index.tsx
@@ -1,0 +1,3 @@
+import TrashButton from './TrashButton'
+
+export default TrashButton
