+++ practice/my-diary/src/index.tsx
@@ -3,7 +3,7 @@ import './index.css'
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import App from './components/app'
+import App from './app/src'
 
 ReactDOM.render(
   <React.StrictMode>
